+++ src/web/assets/src/vue/address-settings/address-settings.vue
@@ -29,9 +29,25 @@ export default {
         LivePreview,
         SubfieldManager,
     },
+    props: {
+        namespace: Object,
+        settings: Object,
+        data: Object,
+        images: Object
+    },
     computed: {
         // Load Pinia store
         ...mapStores(useAddressSettingsStore)
+    },
+    setup(props) {
+        // Get the Pinia store
+        const addressSettingsStore = useAddressSettingsStore();
+
+        // Set Pinia values from props
+        addressSettingsStore.namespace = props.namespace;
+        addressSettingsStore.settings = props.settings;
+        addressSettingsStore.data = props.data;
+        addressSettingsStore.images = props.images;
     },
 }
 </script>