+++ src/web/assets/src/vue/address/address.vue
@@ -51,65 +51,6 @@ export default {
 
         // Whether to show the map by default
         addressStore.showMap = props.settings.showMap;
-    },
-    mounted() {
-        // Get the Pinia store
-        const addressStore = useAddressStore();
-
-        // Get address data & coordinates
-        const addressData = addressStore.data.address;
-        const addressCoords = addressStore.data.coords;
-
-        // Get existing autofill field set
-        const autofill = document.querySelector('mapbox-address-autofill');
-
-        // When a result is selected
-        autofill.addEventListener('retrieve', event => {
-
-            // Reset address meta data
-            addressData.formatted = null;
-            addressData.raw = null;
-
-            // Reset address coordinates
-            addressCoords.lng = null;
-            addressCoords.lat = null;
-
-            // If not already set, reset zoom level
-            addressCoords.zoom = (addressCoords.zoom ?? 11);
-
-            // If results are invalid, bail
-            if (!event || !event.detail || !event.detail.features) {
-                return;
-            }
-
-            // Get feature info
-            const feature = event.detail.features[0] || null;
-
-            // If no feature info, bail
-            if (!feature) {
-                return;
-            }
-
-            // Get coordinates of selected location
-            const lng = (feature.geometry.coordinates[0] ?? null);
-            const lat = (feature.geometry.coordinates[1] ?? null);
-
-            // Update address coordinates
-            addressCoords.lng = (lng ? parseFloat(lng.toFixed(7)) : null);
-            addressCoords.lat = (lat ? parseFloat(lat.toFixed(7)) : null);
-
-            // Update address meta data
-            addressData.formatted = (feature.properties.full_address ?? null);
-            addressData.raw = JSON.stringify(feature);
-
-            // If not changing the map visibility, bail
-            if ('noChange' === addressStore.settings.mapOnSearch) {
-                return;
-            }
-
-            // Change map visibility based on settings
-            addressStore.settings.showMap = ('open' === addressStore.settings.mapOnSearch);
-        });
     }
 }
 </script>